dts: overlay: Fix SPI NOR flash size unit
The `size` property in the `jedec,spi-nor` binding expects the flash
capacity in *Mbits*, not in MiB.

The S25FL256L device is a 256 Mbit flash (32 MiB), but the DTS was
incorrectly using `<32>`, which Zephyr interpreted as 32 Mbit (4 MiB).

Update the `size` property from `<32>` to `<256>` to match the datasheet
and binding requirements.

Signed-off-by: Gaetan Perrot <gaetan.perrot@spacecubics.com>